consolidate redis renderer
