+++ pkg/renderers/redisv1alpha3/render_test.go
@@ -6,9 +6,12 @@
 package redisv1alpha3
 
 import (
+	"context"
 	"testing"
 
+	"github.com/go-logr/logr"
 	"github.com/project-radius/radius/pkg/handlers"
+	"github.com/project-radius/radius/pkg/radlogger"
 	"github.com/project-radius/radius/pkg/radrp/outputresource"
 	"github.com/project-radius/radius/pkg/renderers"
 	"github.com/project-radius/radius/pkg/resourcekinds"
@@ -20,9 +23,18 @@ const (
 	resourceName    = "test-redis"
 )
 
-func Test_Azure_Render_Success(t *testing.T) {
+func createContext(t *testing.T) context.Context {
+	logger, err := radlogger.NewTestLogger(t)
+	if err != nil {
+		t.Log("Unable to initialize logger")
+		return context.Background()
+	}
+	return logr.NewContext(context.Background(), logger)
+}
+
+func Test_Render_Success(t *testing.T) {
 	ctx := createContext(t)
-	renderer := AzureRenderer{}
+	renderer := Renderer{}
 
 	resource := renderers.RendererResource{
 		ApplicationName: applicationName,
@@ -66,9 +78,9 @@ func Test_Azure_Render_Success(t *testing.T) {
 	require.Equal(t, "listKeys", output.SecretValues[renderers.PasswordStringHolder].Action)
 }
 
-func Test_Azure_Render_User_Secrets(t *testing.T) {
+func Test_Render_User_Secrets(t *testing.T) {
 	ctx := createContext(t)
-	renderer := AzureRenderer{}
+	renderer := Renderer{}
 
 	resource := renderers.RendererResource{
 		ApplicationName: applicationName,
@@ -94,9 +106,9 @@ func Test_Azure_Render_User_Secrets(t *testing.T) {
 	require.Equal(t, expectedSecretValues, output.SecretValues)
 }
 
-func Test_Azure_Render_MissingResource(t *testing.T) {
+func Test_Render_NoResourceSpecified(t *testing.T) {
 	ctx := createContext(t)
-	renderer := AzureRenderer{}
+	renderer := Renderer{}
 
 	resource := renderers.RendererResource{
 		ApplicationName: applicationName,
@@ -105,14 +117,14 @@ func Test_Azure_Render_MissingResource(t *testing.T) {
 		Definition:      map[string]interface{}{},
 	}
 
-	_, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: map[string]renderers.RendererDependency{}})
-	require.Error(t, err)
-	require.Equal(t, renderers.ErrResourceMissingForResource.Error(), err.Error())
+	rendererOutput, err := renderer.Render(ctx, renderers.RenderOptions{Resource: resource, Dependencies: map[string]renderers.RendererDependency{}})
+	require.NoError(t, err)
+	require.Equal(t, 0, len(rendererOutput.Resources))
 }
 
-func Test_Azure_Render_InvalidResourceType(t *testing.T) {
+func Test_Render_InvalidResourceType(t *testing.T) {
 	ctx := createContext(t)
-	renderer := AzureRenderer{}
+	renderer := Renderer{}
 
 	resource := renderers.RendererResource{
 		ApplicationName: applicationName,
